Bug fix on redeploy 🐞

+++ pysura/cli/google_root.py
@@ -2833,19 +2833,20 @@ async def SNAKE(_: Request,
         :param memory_default: 2Gi
         :param max_instances_default: 10
         :param default_init: Internal! Do not use!
+        Example: deploy_microservice default 600s 2Gi 10
         """
         if isinstance(default_init, str) and default_init.strip() == "":
             default_init = False
         if isinstance(microservice_name, str) and len(microservice_name.strip()) == 0:
             microservice_name = "default"
-            if isinstance(url_wrapper, str) and len(url_wrapper.strip()) == 0:
-                url_wrapper = "{{HASURA_MICROSERVICE_URL}}"
-            if isinstance(timeout_default, str) and len(timeout_default.strip()) == 0:
-                timeout_default = "600s"
-            if isinstance(memory_default, str) and len(memory_default.strip()) == 0:
-                memory_default = "2Gi"
-            if isinstance(max_instances_default, str) and len(max_instances_default.strip()) == 0:
-                max_instances_default = "10"
+        if isinstance(url_wrapper, str) and len(url_wrapper.strip()) == 0:
+            url_wrapper = "{{HASURA_MICROSERVICE_URL}}"
+        if isinstance(timeout_default, str) and len(timeout_default.strip()) == 0:
+            timeout_default = "600s"
+        if isinstance(memory_default, str) and len(memory_default.strip()) == 0:
+            memory_default = "2Gi"
+        if isinstance(max_instances_default, str) and len(max_instances_default.strip()) == 0:
+            max_instances_default = "10"
         env = self.get_env()
         if env.auth_service_account is None or env.auth_service_account.key_file is None:
             self.log("No auth service account specified", level=logging.ERROR)
@@ -2885,6 +2886,8 @@ async def SNAKE(_: Request,
                                 default_actions.append(f.replace(".py", ""))
                             if microservice_name != "default":
                                 continue
+                            if os.path.isfile(os.path.join(dir_path, f)):
+                                continue
                             shutil.copy(os.path.join(root, f), dir_path)
                     elif "crons" in root:
                         if f != "cron_template.py":
@@ -2895,6 +2898,8 @@ async def SNAKE(_: Request,
                                 default_crons.append(f.replace(".py", ""))
                             if microservice_name != "default":
                                 continue
+                            if os.path.isfile(os.path.join(dir_path, f)):
+                                continue
                             shutil.copy(os.path.join(root, f), dir_path)
                     elif "events" in root:
                         if f != "event_template.py":
@@ -2904,6 +2909,8 @@ async def SNAKE(_: Request,
                             if f != "__init__.py" and microservice_name == "default":
                                 default_events.append(f.replace(".py", ""))
                             if microservice_name != "default":
+                                continue
+                            if os.path.isfile(os.path.join(dir_path, f)):
                                 continue
                             shutil.copy(os.path.join(root, f), dir_path)
         if microservice_name == "default" and default_init:
